Functional rename to coveo-stew
