+++ coveo-stew/coveo_stew/__main__.py
