+++ coveo-stew/coveo_stew/ci/black_runner.py
