+++ coveo-stew/coveo_stew/ci/config.py
