+++ coveo-stew/coveo_stew/ci/mypy_runner.py
