+++ coveo-stew/coveo_stew/ci/reporting.py
