+++ coveo-stew/coveo_stew/environment.py
