+++ coveo-stew/coveo_stew/exceptions.py
