+++ coveo-stew/coveo_stew/metadata/pyproject_api.py
