+++ coveo-stew/coveo_stew/metadata/python_api.py
