+++ coveo-stew/coveo_stew/package_resources/mypy.ini
