+++ coveo-stew/coveo_stew/pyproject.py
