+++ coveo-stew/coveo_stew/utils.py
