+++ coveo-stew/pyproject.toml
