+++ coveo-stew/test_coveo_pyproject/conftest.py
