+++ coveo-stew/test_coveo_pyproject/pyprojet_mock/mock-pyproject-dependency/poetry.toml
